SRSNV report update (#199)
* WIP srsnv report

Major update srsnv report.

Also, allow running on prepared data and trained model (only create
report).

* Add shap package for srsnv report

* "single_read_snv.applicationQC.h5" suffix

* WIP srsnv report fix test bugs

* ppmSeqStrandVcfAnnotator now handles v7 too

* fix bug with v7 ppmSeq tag annotation

* SRSNV training reserve one chrom for test set

When using cross validation, chromosomes are partitioned into training folds.
Now, a configurable number of chromosomes (by default 1)
is excluded from all folds, and reserved exclusively for the test set.

* SRSNV report add training metrics

Add ML_qual/logit histograms, roc auc scores, variance among k-fold CV models.

Cleanup: remove unused plots from report.

* SRSNV report reorder and add training metrics

And also is_mixed + ML_qual in LoD_plot

* SRSNV report fix auc bug and code cleanup

ROC AUC can only be calculated when the dataset is not empty,
and when both classes (True, False) are represented in the dataset.
This commit adds checks for these conditions before calculation
or ROC AUC. It also adds check whether any reads are mixed before
calculating mixed-only statistics.

In addition, the code for SRSNV report is cleaned from unnecessary
code belonging to old report.

* SRSNV report adapt unit tests, small bug fixes

* SRSNV report update plotting utils

* SRSNV update applicationQC.h5 file

* SRSNV report roc auc bug fix

When calculating roc auc on holdout set, model predictions
may be empty (e.g, for is_mixed set when there are no mixed reads).
In this case, an error was raised when calling predict_proba(...)[:,1] (for
the probabilities of the 1 class). This is now fixed.

---------

+++ ugvc/mrd/srsnv_inference_utils.py
@@ -186,7 +186,7 @@ class MLQualAnnotator(VcfAnnotator):
         elif self.chrom_folds is None:
             df[FOLD_ID] = np.nan  # For the case where fold-splitting was done randomly
         else:
-            df[FOLD_ID] = df["chrom"].map(self.chrom_folds, na_action="ignore")
+            df[FOLD_ID] = df["chrom"].map(self.chrom_folds, na_action="ignore").astype('Int64')
         predicted_probability = k_fold_predict_proba(
             self.models, df, features_for_model, self.num_folds, kfold_col=FOLD_ID
         ).values
